% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_diagnostics.R
\name{compute_log_likelihood_Z_ggp}
\alias{compute_log_likelihood_Z_ggp}
\title{Computes log-likelihood of data given current w under GGP model}
\usage{
compute_log_likelihood_Z_ggp(Z, w)
}
\arguments{
\item{Z:}{scalar}

\item{w:}{vector of scalars}
}
\description{
Computes log-likelihood of data given current w under GGP model
}
\keyword{internal}
