% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_diagnostics.R
\name{compute_log_posterior_w}
\alias{compute_log_posterior_w}
\title{Computes log-posterior of log w given current GGP parameters values}
\usage{
compute_log_posterior_w(M_counts, w, w_rem, sigma, tau)
}
\arguments{
\item{M_counts}{matrix with number of edges in every community (TGGP) or vector with total number of edges (GGP), for each node}

\item{w:}{vector of positive scalars}

\item{w_rem:}{positive scalar; total mass of discarded nodes (previous sample)}

\item{sigma:}{real in (-Inf, 1) (previous sample or fixed)}

\item{tau:}{positive scalar (previous sample or fixed)}
}
\description{
Computes log-posterior of log w given current GGP parameters values
}
\keyword{internal}
