% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ggp_mcmc.R
\name{fit_ggp_mcmc}
\alias{fit_ggp_mcmc}
\title{Fit Generalized Gamma Process to a simple undirected network via Gibbs Sampling 
following the method described in Caron and Fox (2017)}
\usage{
fit_ggp_mcmc(Z, mcmc_settings, general_settings, ggp_param = list())
}
\arguments{
\item{Z}{sparse symmetric binary adjacency matrix of undirected network.}

\item{mcmc_settings}{list providing desired mcmc settings (e.g. number of iterations). See Details.}

\item{general_settings}{list indicating logistics details (e.g. path for saving). See Details.}

\item{ggp_param}{list (optional) can include fields alpha, sigma and tau}
}
\description{
ADD DESCRIPTION HERE
}
\examples{
\dontrun{
# Fit TGGP model on a simulated network with 3 communities of similar size
tggp_network <- sample_tggp_network(alpha = 50, sigma = 0.5, tau = 1, K = 3)
Z <- tggp_network$Z
mcmc_settings <- list(n_iter = 10)
general_settings <- list(path_for_saving = "experiments/results"),
ggp_param <- list(K = 5)
Z_fit <- fit_tggp_mcmc(Z, mcmc_settings, general_settings, ggp_param)
}

}
\references{
F. Ricci, M. Guindani, and E. Sudderth (2022). Thinned Random Measures for Sparse Graphs with Overlapping Communities. Neural Information Processing Systems 

Caron, F., & Fox, E. B. (2017). Sparse graphs using exchangeable random measures. Journal of the Royal Statistical Society. Series B, Statistical Methodology, 79(5), 12
}
