% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tggp_mcmc.R
\name{fit_tggp_mcmc}
\alias{fit_tggp_mcmc}
\title{Fit Thinned Generalized Gamma Process to a simple undirected network via Gibbs Sampling}
\usage{
fit_tggp_mcmc(Z, mcmc_settings, general_settings, tggp_param)
}
\arguments{
\item{Z}{sparse symmetric binary adjacency matrix of undirected network.}

\item{mcmc_settings}{list specifying desired number of iterations and (optionally) other mcmc settings. See Details.}

\item{general_settings}{list specifying path to store results and (optionally) other general settings. See Details.}

\item{tggp_param}{list specifying desired maximum number of communities and (optionally) other hyperparameters (only if user wishes to fix them or provide initial values). See Details.}
}
\value{
\subsection{A list with the following components: } {
   \describe{
     \item{mcmc_samples} {MCMC samples of model parameters (nodes' sociabilities and community memberships) and hyperparameters (relative size of communities, GGP parameters)).}
     \item{mcmc_acceptance} {Hamiltonian Monte Carlo and Metropolis Hastings acceptance rates.}
   }
}
}
\description{
ADD DESCRIPTION HERE
}
\details{
ADD DETAILS ON ARGUMENTS' NAMED FIELDS
}
\examples{
\dontrun{
# Fit TGGP model on a simulated network with 3 communities of similar size
tggp_network <- sample_tggp_network(alpha = 50, sigma = 0.5, tau = 1, K = 3)
Z <- tggp_network$Z
mcmc_settings <- list(n_iter = 10)
general_settings <- list(path_for_saving = "experiments/results"),
model_hyper <- list(K = 5)
Z_fit <- fit_tggp_mcmc(Z, mcmc_settings, general_settings, model_hyper)
}

}
\references{
F. Ricci, M. Guindani, and E. Sudderth (2022). Thinned Random Measures for Sparse Graphs with Overlapping Communities. Neural Information Processing Systems 

Caron, F., & Fox, E. B. (2017). Sparse graphs using exchangeable random measures. Journal of the Royal Statistical Society. Series B, Statistical Methodology, 79(5), 12
}
