% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hyperparam.R
\name{gen_U}
\alias{gen_U}
\title{Utility function for r_et_stable (generates U)}
\usage{
gen_U(w1, w2, w3, gam)
}
\arguments{
\item{w1:}{scalar}

\item{w2:}{scalar}

\item{w3:}{scalar}

\item{gam:}{scalar}
}
\description{
Utility function for r_et_stable (generates U)
}
\keyword{internal}
