% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_w.R
\name{grad_loglik_log_w}
\alias{grad_loglik_log_w}
\title{Compute gradient of negative log-likelihood of log w ----------------------------}
\usage{
grad_loglik_log_w(N_counts, w, sigma, tau, w_rem)
}
\arguments{
\item{N_counts:}{vector of positive integers, nodes degrees (previous sample, or observed)}

\item{w:}{vector of positive scalars, nodes sociability (previous sample)}

\item{sigma:}{real in (-Inf, 1) (previous sample or fixed)}

\item{tau:}{positive scalar (previous sample or fixed)}

\item{w_rem:}{positive scalar; total mass of discarded nodes (previous sample)}
}
\description{
Compute gradient of negative log-likelihood of log w ----------------------------
}
\references{
Adapted to R from original Matlab code by F. Caron:
Copyright (C) Francois Caron, University of Oxford
see: http://www.stats.ox.ac.uk/~caron/code/bnpgraph/index.html
Caron, F., & Fox, E. B. (2017). Sparse graphs using exchangeable random measures. Journal of the Royal Statistical Society. Series B, Statistical Methodology, 79(5), 12
}
\keyword{internal}
