% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_utils.R
\name{initialize_model_param}
\alias{initialize_model_param}
\title{Initializes model-related stuff (and for fit_ggp_mcmc, initialize hyperparams)}
\usage{
initialize_model_param(model_param, N, model = "tggp")
}
\arguments{
\item{model_param}{list; if ``model = tggp`` then the list needs to include a field K specifying the upper bound to the number of communities}

\item{N}{int; number of nodes in observed network}

\item{model}{string; defaults to "tggp". If given a different value, only hyperparameters from GGP model are initialized}
}
\description{
Initializes model-related stuff (and for fit_ggp_mcmc, initialize hyperparams)
}
\keyword{internal}
