% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_utils.R
\name{initialize_tggp}
\alias{initialize_tggp}
\title{Initializes all hyperparameters and parameters for fit_tggp_mcmc}
\usage{
initialize_tggp(Z, n_iter_initial, path_for_saving, tggp_param, verbose)
}
\arguments{
\item{Z}{sparse symmetric binary adjacency matrix of undirected network.}

\item{n_iter_initial}{int; desired number of GGP mcmc runs}

\item{path_for_saving}{directory; path to store results.}
}
\description{
Initializes all hyperparameters and parameters for fit_tggp_mcmc
}
\keyword{internal}
