% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_utils.R
\name{initialize_with_ggp}
\alias{initialize_with_ggp}
\title{Initializes nodes sociabilities and GGP hyperparameters running GGP mcmc}
\usage{
initialize_with_ggp(Z, n_iter_initial, path_for_saving)
}
\arguments{
\item{Z}{sparse symmetric binary adjacency matrix of undirected network.}

\item{n_iter_initial}{int; desired number of GGP mcmc runs}

\item{path_for_saving}{directory; path to store results.}
}
\description{
Initializes nodes sociabilities and GGP hyperparameters running GGP mcmc
}
\keyword{internal}
