% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hyperparam.R
\name{psi_GGP}
\alias{psi_GGP}
\title{Laplace exponent of GGP: Psi function}
\usage{
psi_GGP(t, alpha, sigma, tau)
}
\arguments{
\item{t:}{value at which to evaluate the Laplace exponent}

\item{alpha:}{positive scalar; truncation of GGP (previous sample or fixed)}

\item{sigma:}{real in (-Inf, 1) (previous sample or fixed)}

\item{tau:}{positive scalar (previous sample or fixed)}
}
\description{
psi(t) = -log ( E[exp(-t * sum_i w_i)] )
       = alpha/sigma * ( (t+tau)^sigma - tau^sigma))
where
(w_i)_{i=1,2,..} are the points of a Poisson process on R_+ of mean measure 
rho(dw) = alpha/Gamma(1-sigma) * w^{-1-sigma} * exp(-tau*w)dw
}
\references{
Adapted to R from original Matlab code by F. Caron:
Copyright (C) Francois Caron, University of Oxford
see: http://www.stats.ox.ac.uk/~caron/code/bnpgraph/index.html
Caron, F., & Fox, E. B. (2017). Sparse graphs using exchangeable random measures. Journal of the Royal Statistical Society. Series B, Statistical Methodology, 79(5), 12
}
\keyword{internal}
