% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hyperparam.R
\name{r_et_stable}
\alias{r_et_stable}
\title{Sample from from the exponentially tilted stable distribution
distributed fron an exponentially tilted stable distribution 
with Laplace transform (in z) exp(-V0 * ((z + tau)^alpha - tau^alpha))
Uses the algorithm proposed in (Devroye, 2009) with corrections pointed 
out in (Hofert, 2011)}
\usage{
r_et_stable(V0, alpha, tau, n = 1)
}
\arguments{
\item{V0:}{positive scalar}

\item{alpha:}{positive scalar; truncation of GGP (previous sample or fixed)}

\item{tau:}{positive scalar (previous sample or fixed)}

\item{n:}{integer}
}
\description{
Sample from from the exponentially tilted stable distribution
distributed fron an exponentially tilted stable distribution 
with Laplace transform (in z) exp(-V0 * ((z + tau)^alpha - tau^alpha))
Uses the algorithm proposed in (Devroye, 2009) with corrections pointed 
out in (Hofert, 2011)
}
\references{
Adapted to R from original Matlab code by F. Caron:
Copyright (C) Francois Caron, University of Oxford
see: http://www.stats.ox.ac.uk/~caron/code/bnpgraph/index.html
Caron, F., & Fox, E. B. (2017). Sparse graphs using exchangeable random measures. Journal of the Royal Statistical Society. Series B, Statistical Methodology, 79(5), 12
Luc Devroye. Random variate generation for exponentially and polynomially tilted stable distributions. ACM Transactions on Modeling and Computer Simulation, vol. 19(4), 2009.
Marius Hofert. Sampling exponentially tilted stable distributions. ACM Transactions on Modeling and Computer Simulation, vol. 22(1), 2011
}
\keyword{internal}
