% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_ggp.R
\name{rand_ggp}
\alias{rand_ggp}
\title{(Approximate) sample from the Generalized Gamma Process}
\source{
R Code adapted from original Matlab code by F. Caron:
Copyright (C) Francois Caron, University of Oxford.

See: http://www.stats.ox.ac.uk/~caron/code/bnpgraph/index.html
}
\usage{
rand_ggp(alpha, sigma, tau)
}
\arguments{
\item{alpha}{positive real number. Affects the number of nodes and edges (higher alpha leads to larger networks).}

\item{sigma}{real number. in (-Inf, 1). sigma < 0 corresponds to a finite-activity GGP (dense case).}

\item{tau}{positive real number. Affects the tails of the degree distribution (higher tau leads to faster decay).}
}
\value{
A real-valued vector with an approximate sample from the GGP.
}
\description{
Samples points of the Generalized Gamma Process (GGP) with Lévy measure
\deqn{[alpha / Gamma(1-sigma)] * w^(-1-sigma)  * exp(-tau * w)}
When sigma < 0, exact sampling is possible.
For sigma >= 0, approximate samples are obtained using the adaptive 
thinning strategy described in Favaro and Teh (2013).
}
\examples{
w <- simulate_ggp(20, 0.5, 1)

}
\references{
Caron, F., & Fox, E. B. (2017). Sparse graphs using exchangeable random measures. Journal of the Royal Statistical Society. Series B, Statistical Methodology, 79(5), 12

S. Favaro and Y.W. Teh. MCMC for normalized random measure mixture models. Statistical Science, vol.28(3), pp.335-359, 2013.
}
