% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_tggp_network.R
\name{rand_tggp_network}
\alias{rand_tggp_network}
\title{Sample of a network from the Thinned Generalized Gamma Process}
\usage{
rand_tggp_network(
  alpha,
  sigma,
  tau,
  K = 2,
  beta = NULL,
  gamma = NULL,
  zeta = NULL,
  return_thinned = FALSE,
  only_simulate_N = FALSE
)
}
\arguments{
\item{alpha}{positive real number. Affects the number of nodes and edges (higher alpha leads to larger networks).}

\item{sigma}{real number. in (-Inf, 1). sigma < 0 corresponds to a finite-activity GGP (dense case).}

\item{tau}{positive real number. Affects the tails of the degree distribution (higher tau leads to faster decay).}

\item{K}{positive integer. Number of communities that nodes can have memberships in.}

\item{beta}{positive real-valued vector. beta[k] affects the overall size of community k. Defaults to a vector with all elements equal to 1/K. If K is specified, length of beta must be equal to K.}

\item{gamma}{real number. If gamma is specified, beta is sampled from a K-dimensional Dirichlet with concentration parameters gamma/K. If both gamma and beta are provided, provided gamma is ignored.}

\item{zeta}{real number. Node community memberships are sampled from a K-dimensional Dirichlet with concentration parameter (zeta*beta)/K. Defaults to 0.5.}

\item{return_thinned}{logical. Defaults to FALSE. If TRUE, thinned edges, as well as sociabilities and community memberships of nodes with no unthinned edge, are also returned.}

\item{only_simulate_N}{logical. Defaults to FALSE. If TRUE, return only number of unthinned nodes and of thinned nodes.}
}
\value{
\describe{
A list with at least the following components: 
\item{edges}{a matrix storing source node, target node, and community 
  assignment of source and target nodes for each unthinned edge.}
\item{Z}{sparse symmetric binary adjacency matrix of undirected network.}
\item{nodes}{a matrix storing sociability and community memberships of 
  all nodes with at least one unthinned edge.}
If ``return_thinned=TRUE``, the following additional components are also 
included in the returned list:
\item{edges_thinned}{a matrix storing source node, target node, and community 
  assignments of source and target nodes for each thinned edge.}
\item{Z_thinned}{sparse symmetric binary adjacency matrix of thinned edges.}
\item{nodes_thinned}{a matrix storing sociability and community memberships 
  of nodes with no unthinned edge.}
}
}
\description{
ADD DESCRIPTION HERE
}
\examples{
# Sample a TGGP network with 50 communities of similar size
tggp_1 <- sample_tggp_network(200, 0.5, 1, K = 50)


}
\references{
F. Ricci, M. Guindani, and E. Sudderth (2022). Thinned Random Measures for Sparse Graphs with Overlapping Communities. Neural Information Processing Systems 

Caron, F., & Fox, E. B. (2017). Sparse graphs using exchangeable random measures. Journal of the Royal Statistical Society. Series B, Statistical Methodology, 79(5), 12
}
