% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hyperparam.R
\name{ratio_B}
\alias{ratio_B}
\title{Utility function for r_et_stable}
\usage{
ratio_B(x, sigma)
}
\arguments{
\item{x:}{scalar}

\item{sigma:}{real in (-Inf, 1) (previous sample or fixed)}
}
\description{
Utility function for r_et_stable
}
\keyword{internal}
