% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hyperparam.R
\name{rGGPsum}
\alias{rGGPsum}
\title{Sample from distribution of GGP total mass}
\usage{
rGGPsum(alpha, sigma, tau)
}
\arguments{
\item{t:}{value at which to evaluate the Laplace exponent}

\item{alpha:}{positive scalar; truncation of GGP (previous sample or fixed)}

\item{sigma:}{real in (-Inf, 1) (previous sample or fixed)}

\item{tau:}{positive scalar (previous sample or fixed)}
}
\description{
It generates a realization of the random variable S with Laplace transform
E[e^-(t*S)] = exp(-alpha/sigma * [(t+tau)^sigma - tau^sigma])
}
\references{
Adapted to R from original Matlab code by F. Caron:
Copyright (C) Francois Caron, University of Oxford
see: http://www.stats.ox.ac.uk/~caron/code/bnpgraph/index.html
Caron, F., & Fox, E. B. (2017). Sparse graphs using exchangeable random measures. Journal of the Royal Statistical Society. Series B, Statistical Methodology, 79(5), 12
}
\keyword{internal}
