% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_utils.R
\name{set_general_settings}
\alias{set_general_settings}
\title{Sets all general settings}
\usage{
set_general_settings(general_settings, n_iter)
}
\arguments{
\item{general_settings}{list; field "path_for_saving" must indicate desired number of iteration; other fields are optional.}

\item{n_iter}{int; number of mcmc iterations}
}
\description{
Sets all general settings
}
\keyword{internal}
