% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_utils.R
\name{set_mcmc_settings}
\alias{set_mcmc_settings}
\title{Sets all MCMC settings}
\usage{
set_mcmc_settings(mcmc_settings, N)
}
\arguments{
\item{mcmc_settings}{list; field "n_iter" must indicate desired number of iteration; other fields are optional.}

\item{N_tot}{int; number of nodes in observed network}

\item{model}{string; defaults to "tggp". If given a different value, only hyperparameters from GGP model are initialized}
}
\description{
Sets all MCMC settings
}
\keyword{internal}
