% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hyperparam.R
\name{sinc}
\alias{sinc}
\title{Utility function for r_et_stable (computes sin(x)/x)}
\usage{
sinc(x)
}
\arguments{
\item{x:}{scalar}
}
\description{
Utility function for r_et_stable (computes sin(x)/x)
}
\keyword{internal}
