% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hyperparam.R
\name{update_ggp_hyperparam}
\alias{update_ggp_hyperparam}
\title{Update total mass of removed nodes and hyperparameters given the rest,
using MH as described in Caron and Fox, Appendix F.2}
\usage{
update_ggp_hyperparam(w, w_rem, alpha, sigma, tau, ggp_param, rw_alpha = FALSE)
}
\arguments{
\item{w:}{vector of positive scalars, nodes sociability (previous sample)}

\item{w_rem:}{positive scalar; total mass of discarded nodes (previous sample)}

\item{alpha:}{positive scalar; truncation of GGP (previous sample or fixed)}

\item{sigma:}{real in (-Inf, 1) (previous sample or fixed)}

\item{tau:}{positive scalar (previous sample or fixed)}

\item{ggp_param:}{list including information on alpha, sigma and tau}
}
\description{
Update total mass of removed nodes and hyperparameters given the rest,
using MH as described in Caron and Fox, Appendix F.2
}
\references{
Adapted to R from original Matlab code by F. Caron:
Copyright (C) Francois Caron, University of Oxford
see: http://www.stats.ox.ac.uk/~caron/code/bnpgraph/index.html
Caron, F., & Fox, E. B. (2017). Sparse graphs using exchangeable random measures. Journal of the Royal Statistical Society. Series B, Statistical Methodology, 79(5), 12
}
\keyword{internal}
