% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hyperparam.R
\name{update_N_tot_to_N_ratio}
\alias{update_N_tot_to_N_ratio}
\title{Update total mass of removed nodes and hyperparameters given the rest,
using MH as described in Caron and Fox, Appendix F.2}
\usage{
update_N_tot_to_N_ratio(alpha, sigma, tau, beta, num_N_samples)
}
\arguments{
\item{alpha:}{positive scalar; truncation of GGP (previous sample or fixed)}

\item{sigma:}{real in (-Inf, 1) (previous sample or fixed)}

\item{tau:}{positive scalar (previous sample or fixed)}

\item{beta:}{vector of overall frequencies of each K_max communities (previous sample or fixed)}

\item{num_N_tot_samples:}{integer: number of samples to base update on}
}
\description{
Update total mass of removed nodes and hyperparameters given the rest,
using MH as described in Caron and Fox, Appendix F.2
}
\keyword{internal}
