% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_w.R
\name{update_w_new}
\alias{update_w_new}
\title{Update nodes sociability parameters given the rest, using HMC as
described in Caron and Fox, Appendix F.1 a}
\usage{
update_w_new(w, w_rem, sigma, tau, N_counts, L, epsilon)
}
\arguments{
\item{w:}{vector of positive scalars, nodes sociability (previous sample)}

\item{w_rem:}{positive scalar; total mass of discarded nodes (previous sample)}

\item{sigma:}{real in (-Inf, 1) (previous sample or fixed)}

\item{tau:}{positive scalar (previous sample or fixed)}

\item{N_counts:}{vector of positive integers, nodes degrees (previous sample, or observed)}

\item{L:}{HMC parameter, num. of leapfrog steps at each iteration}

\item{epsilon:}{HMC parameter, leapfrog step-size}
}
\description{
Update nodes sociability parameters given the rest, using HMC as
described in Caron and Fox, Appendix F.1 a
}
\references{
Adapted to R from original Matlab code by F. Caron:
Copyright (C) Francois Caron, University of Oxford
see: http://www.stats.ox.ac.uk/~caron/code/bnpgraph/index.html
Caron, F., & Fox, E. B. (2017). Sparse graphs using exchangeable random measures. Journal of the Royal Statistical Society. Series B, Statistical Methodology, 79(5), 12
}
\keyword{internal}
