% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hyperparam.R
\name{zolotarev}
\alias{zolotarev}
\title{Utility function for r_et_stable (evaluates Zolotarev function)}
\usage{
zolotarev(u, sigma)
}
\arguments{
\item{u:}{scalar}

\item{sigma:}{real in (-Inf, 1) (previous sample or fixed)}
}
\description{
Utility function for r_et_stable (evaluates Zolotarev function)
}
\keyword{internal}
